extracted RateService for re-usability
